requirements.txt
dash==2.18.2
dash-bootstrap-components==1.7.1
dill==0.3.9
numpy==2.1.3
soundfile==0.13.1
librosa==0.10.2.post1
pandas==2.2.3
plotly_express==0.4.1
sounddevice==0.5.1
pywebview==5.4
